tests: move experiments fixtures into conftest
